+++ tests/tests_app/runners/test_multiprocess.py
@@ -4,7 +4,6 @@ from unittest import mock
 from unittest.mock import Mock
 
 import pytest
-
 from lightning.app import LightningApp, LightningFlow, LightningWork
 from lightning.app.core import constants
 from lightning.app.frontend import StaticWebFrontend, StreamlitFrontend