+++ tests/tests_pytorch/loops/test_training_epoch_loop.py
@@ -15,7 +15,6 @@ import logging
 from unittest.mock import Mock, patch
 
 import pytest
-
 from lightning.pytorch.demos.boring_classes import BoringModel
 from lightning.pytorch.trainer.trainer import Trainer