+++ tests/tests_pytorch/models/test_hooks.py
@@ -18,11 +18,11 @@ from unittest.mock import ANY, PropertyMock
 
 import pytest
 import torch
+from lightning.pytorch import Callback, LightningDataModule, LightningModule, Trainer, __version__
+from lightning.pytorch.demos.boring_classes import BoringDataModule, BoringModel, RandomDataset
 from torch import Tensor
 from torch.utils.data import DataLoader
 
-from lightning.pytorch import __version__, Callback, LightningDataModule, LightningModule, Trainer
-from lightning.pytorch.demos.boring_classes import BoringDataModule, BoringModel, RandomDataset
 from tests_pytorch.helpers.runif import RunIf