+++ tests/tests_pytorch/utilities/test_grads.py
@@ -16,7 +16,6 @@ from unittest.mock import Mock
 import pytest
 import torch
 import torch.nn as nn
-
 from lightning.pytorch.utilities import grad_norm